+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+next
+=====
+
+  * Supported running empty migration scripts.
+
 1.0.0
 =====